refactor: 각 필드들 features 레이어로 분리
