+++ apps/web/src/features/auth/ui/enrollment-document-upload-field/index.ts
@@ -1,0 +1,1 @@
+export * from './EnrollmentDocumentUploadField';
