+++ apps/web/src/features/auth/ui/enrollment-verification-enrollment-year-field/index.ts
@@ -1,0 +1,1 @@
+export * from './EnrollmentVerificationEnrollmentYearField';
