+++ apps/web/src/features/auth/ui/enrollment-verification-name-field/index.ts
@@ -1,0 +1,1 @@
+export * from './EnrollmentVerificationNameField';
