--- apps/web/src/features/auth/ui/enrollment-verification-section/index.ts
@@ -1,3 +1,0 @@
-export * from './BaseInfoSection';
-export * from './AcademicInfoSection';
-export * from './DocumentSection';
