+++ apps/web/src/features/auth/ui/index.ts
@@ -2,8 +2,14 @@ export { KakaoLoginButton } from './kakao-login-button';
 export { GoogleLoginButton } from './google-login-button';
 export { AppleLoginButton } from './apple-login-button';
 export { EmailLoginButton } from './email-login-button';
-export * from './enrollment-verification-section';
 export * from './enrollment-step-action-button';
+export * from './enrollment-document-upload-field';
+export * from './enrollment-verification-enrollment-year-field';
+export * from './enrollment-verification-graduation-year-field';
+export * from './enrollment-verification-major-field';
+export * from './enrollment-verification-name-field';
+export * from './enrollment-verification-state-field';
+export * from './enrollment-verification-student-id-field';
 export * from './oauth-additional-info-name-field';
 export * from './oauth-additional-info-phone-number-field';
 export * from './oauth-additional-info-nickname-field';
